add standard, split out types of "reduce" tests
