+++ test/graph-tools/reduce/merge.test.js
@@ -1,7 +1,7 @@
 const test = require('tape')
-const reduce = require('../../graph-tools/reduce')
-const OverwriteStrategy = require('../../strategy/overwrite')
-const Strategy = require('../../strategy/compose')
+const reduce = require('../../../graph-tools/reduce')
+const OverwriteStrategy = require('../../../strategy/overwrite')
+const Strategy = require('../../../strategy/compose')
 
 const StringAppendStrategy = () => ({ // stringAppend
   concat: (a, b) => a + b,
@@ -15,11 +15,6 @@ const strategy = Strategy({
   title: OverwriteStrategy(),
   ouji: StringAppendStrategy()
 })
-// TODO change reduce to take compose(composition)
-
-function titleTransformation (val) {
-  return { set: val }
-}
 
 test('reduce', t => {
   //     A   (root)
@@ -29,7 +24,7 @@ test('reduce', t => {
   const A = {
     key: 'A',
     thread: { root: null, previous: null },
-    title: titleTransformation('my root message'),
+    title: { set: 'my root message' },
     ouji: 'hello'
   }
   const B = {
@@ -40,7 +35,7 @@ test('reduce', t => {
   const C = {
     key: 'C',
     thread: { root: 'A', previous: ['A'] },
-    title: titleTransformation('edited message'),
+    title: { set: 'edited message' },
     ouji: ' world'
   }
 
@@ -48,11 +43,11 @@ test('reduce', t => {
     reduce(A, [B, C], strategy),
     {
       B: {
-        title: titleTransformation('my root message'),
+        title: { set: 'my root message' },
         ouji: 'hello mix'
       },
       C: {
-        title: titleTransformation('edited message'),
+        title: { set: 'edited message' },
         ouji: 'hello world'
       }
     },
@@ -68,7 +63,7 @@ test('reduce', t => {
   const D = {
     key: 'D',
     thread: { root: 'A', previous: ['B', 'C'] },
-    title: titleTransformation('edited and merged!'),
+    title: { set: 'edited and merged!' },
     ouji: 'hello world (mix)'
   }
   // this is a valid merge because it resolves the conflict present between B + C
@@ -78,89 +73,11 @@ test('reduce', t => {
     reduce(A, [B, C, D], strategy),
     {
       D: {
-        title: titleTransformation('edited and merged!'),
+        title: { set: 'edited and merged!' },
         ouji: 'hello world (mix)'
       }
     },
     'merge with conflict'
-  )
-
-  t.end()
-})
-
-test('reduce (custom getThread)', t => {
-  //   A   (root)
-  //   |
-  //   B
-
-  const A = {
-    key: 'A',
-    tangles: {
-      conversation: [null, null]
-    },
-    title: titleTransformation('my root message'),
-    ouji: 'hello'
-  }
-  const B = {
-    key: 'B',
-    tangles: {
-      conversation: ['A', ['A']]
-    },
-    ouji: ' mix'
-  }
-
-  const getThread = node => {
-    const [ root, previous ] = node.tangles.conversation
-
-    return { root, previous }
-  }
-
-  t.deepEqual(
-    reduce(A, [B], strategy, { getThread }),
-    {
-      B: {
-        title: titleTransformation('my root message'),
-        ouji: 'hello mix'
-      }
-    },
-    'custome getThread works'
-  )
-
-  t.end()
-})
-
-test('reduce (custom getTransformation)', t => {
-  //   A   (root)
-  //   |
-  //   B
-
-  const A = {
-    key: 'A',
-    thread: { root: null, previous: null },
-    mutations: {
-      title: titleTransformation('my root message'),
-      ouji: 'hello'
-    }
-  }
-  const B = {
-    key: 'B',
-    thread: { root: 'A', previous: ['A'] },
-    mutations: {
-      ouji: ' mix'
-    }
-  }
-
-  const getTransformation = node => node.mutations
-
-  t.deepEqual(
-    reduce(A, [B], strategy, { getTransformation }),
-    {
-      B: {
-        title: titleTransformation('my root message'),
-        ouji: 'hello mix'
-      }
-    },
-    'custom getTransformation works'
   )
 
   t.end()
@@ -176,7 +93,7 @@ test('reduce (invalid merge)', { todo: true }, t => {
   const A = {
     key: 'A',
     thread: { root: null, previous: null },
-    title: titleTransformation('my root message'),
+    title: { set: 'my root message' },
     ouji: 'hello'
   }
   const B = {
@@ -187,13 +104,13 @@ test('reduce (invalid merge)', { todo: true }, t => {
   const C = {
     key: 'C',
     thread: { root: 'A', previous: ['A'] },
-    title: titleTransformation('edited message'),
+    title: { set: 'edited message' },
     ouji: ' world'
   }
   const Dud = {
     key: 'D',
     thread: { root: 'A', previous: ['B', 'C'] },
-    title: titleTransformation('edited and merged!')
+    title: { set: 'edited and merged!' }
     // ouji: identity()
   }
   // this is an invalid merge message because Dud fails to resolve conflict between B + C
@@ -203,11 +120,11 @@ test('reduce (invalid merge)', { todo: true }, t => {
     reduce(A, [B, C, Dud], strategy),
     {
       B: {
-        title: titleTransformation('my root message'),
+        title: { set: 'my root message' },
         ouji: 'hello mix'
       },
       C: {
-        title: titleTransformation('edited message'),
+        title: { set: 'edited message' },
         ouji: 'hello world'
       }
     },
@@ -229,25 +146,25 @@ test('reduce (automerge)', { todo: true }, t => {
   const A = {
     key: 'A',
     thread: { root: null, previous: null },
-    title: titleTransformation('my root message'),
+    title: { set: 'my root message' },
     ouji: 'hello'
   }
   const B = {
     key: 'B',
     thread: { root: 'A', previous: ['A'] },
-    title: titleTransformation('one two'),
+    title: { set: 'one two' },
     ouji: ' mix'
   }
   const C = {
     key: 'C',
     thread: { root: 'A', previous: ['A'] },
-    title: titleTransformation('one'),
+    title: { set: 'one' },
     ouji: ' world'
   }
   const D = {
     key: 'D',
     thread: { root: 'A', previous: ['C'] },
-    title: titleTransformation('one two')
+    title: { set: 'one two' }
   }
   const E = {
     key: 'E',
@@ -262,7 +179,7 @@ test('reduce (automerge)', { todo: true }, t => {
     reduce(A, [B, C, D, E], strategy),
     {
       E: {
-        title: titleTransformation('my root message'),
+        title: { set: 'my root message' },
         ouji: 'hello mix-world!'
       }
     },
